Extract MatchExpressionTreeVisitor class
