+++ src/main/java/io/cryostat/rules/MatchExpressionValidator.java
@@ -37,211 +37,15 @@
  */
 package io.cryostat.rules;
 
-import jdk.nashorn.api.tree.BreakTree;
-import jdk.nashorn.api.tree.CaseTree;
-import jdk.nashorn.api.tree.CatchTree;
-import jdk.nashorn.api.tree.ClassDeclarationTree;
-import jdk.nashorn.api.tree.ClassExpressionTree;
 import jdk.nashorn.api.tree.CompilationUnitTree;
-import jdk.nashorn.api.tree.ContinueTree;
-import jdk.nashorn.api.tree.DebuggerTree;
-import jdk.nashorn.api.tree.DoWhileLoopTree;
-import jdk.nashorn.api.tree.ErroneousTree;
-import jdk.nashorn.api.tree.ExportEntryTree;
-import jdk.nashorn.api.tree.ForInLoopTree;
-import jdk.nashorn.api.tree.ForLoopTree;
-import jdk.nashorn.api.tree.ForOfLoopTree;
-import jdk.nashorn.api.tree.FunctionCallTree;
-import jdk.nashorn.api.tree.FunctionDeclarationTree;
-import jdk.nashorn.api.tree.FunctionExpressionTree;
-import jdk.nashorn.api.tree.ImportEntryTree;
-import jdk.nashorn.api.tree.InstanceOfTree;
-import jdk.nashorn.api.tree.LabeledStatementTree;
-import jdk.nashorn.api.tree.ModuleTree;
-import jdk.nashorn.api.tree.NewTree;
 import jdk.nashorn.api.tree.Parser;
-import jdk.nashorn.api.tree.RegExpLiteralTree;
-import jdk.nashorn.api.tree.ReturnTree;
-import jdk.nashorn.api.tree.SimpleTreeVisitorES5_1;
-import jdk.nashorn.api.tree.SpreadTree;
-import jdk.nashorn.api.tree.ThrowTree;
-import jdk.nashorn.api.tree.Tree;
 import jdk.nashorn.api.tree.TreeVisitor;
-import jdk.nashorn.api.tree.TryTree;
-import jdk.nashorn.api.tree.UnaryTree;
-import jdk.nashorn.api.tree.WhileLoopTree;
-import jdk.nashorn.api.tree.WithTree;
-import jdk.nashorn.api.tree.YieldTree;
 
 public class MatchExpressionValidator {
 
     private final Parser parser = Parser.create();
 
-    private final TreeVisitor<Void, String> treeVisitor =
-            new SimpleTreeVisitorES5_1<Void, String>() {
-                Void fail(Tree node, String matchExpression) {
-                    throw new IllegalMatchExpressionException(node, matchExpression);
-                }
-
-                @Override
-                public Void visitBreak(BreakTree node, String matchExpression) {
-                    return fail(node, matchExpression);
-                }
-
-                @Override
-                public Void visitCase(CaseTree node, String matchExpression) {
-                    return fail(node, matchExpression);
-                }
-
-                @Override
-                public Void visitCatch(CatchTree node, String matchExpression) {
-                    return fail(node, matchExpression);
-                }
-
-                @Override
-                public Void visitClassDeclaration(
-                        ClassDeclarationTree node, String matchExpression) {
-                    return fail(node, matchExpression);
-                }
-
-                @Override
-                public Void visitClassExpression(ClassExpressionTree node, String matchExpression) {
-                    return fail(node, matchExpression);
-                }
-
-                @Override
-                public Void visitDoWhileLoop(DoWhileLoopTree node, String matchExpression) {
-                    return fail(node, matchExpression);
-                }
-
-                @Override
-                public Void visitErroneous(ErroneousTree node, String matchExpression) {
-                    return fail(node, matchExpression);
-                }
-
-                @Override
-                public Void visitExportEntry(ExportEntryTree node, String matchExpression) {
-                    return fail(node, matchExpression);
-                }
-
-                @Override
-                public Void visitForInLoop(ForInLoopTree node, String matchExpression) {
-                    return fail(node, matchExpression);
-                }
-
-                @Override
-                public Void visitForLoop(ForLoopTree node, String matchExpression) {
-                    return fail(node, matchExpression);
-                }
-
-                @Override
-                public Void visitForOfLoop(ForOfLoopTree node, String matchExpression) {
-                    return fail(node, matchExpression);
-                }
-
-                @Override
-                public Void visitFunctionCall(FunctionCallTree node, String matchExpression) {
-                    return fail(node, matchExpression);
-                }
-
-                @Override
-                public Void visitFunctionDeclaration(
-                        FunctionDeclarationTree node, String matchExpression) {
-                    return fail(node, matchExpression);
-                }
-
-                @Override
-                public Void visitFunctionExpression(
-                        FunctionExpressionTree node, String matchExpression) {
-                    return fail(node, matchExpression);
-                }
-
-                @Override
-                public Void visitImportEntry(ImportEntryTree node, String matchExpression) {
-                    return fail(node, matchExpression);
-                }
-
-                @Override
-                public Void visitInstanceOf(InstanceOfTree node, String matchExpression) {
-                    return fail(node, matchExpression);
-                }
-
-                @Override
-                public Void visitLabeledStatement(
-                        LabeledStatementTree node, String matchExpression) {
-                    return fail(node, matchExpression);
-                }
-
-                @Override
-                public Void visitModule(ModuleTree node, String matchExpression) {
-                    return fail(node, matchExpression);
-                }
-
-                @Override
-                public Void visitNew(NewTree node, String matchExpression) {
-                    return fail(node, matchExpression);
-                }
-
-                @Override
-                public Void visitRegExpLiteral(RegExpLiteralTree node, String matchExpression) {
-                    return fail(node, matchExpression);
-                }
-
-                @Override
-                public Void visitReturn(ReturnTree node, String matchExpression) {
-                    return fail(node, matchExpression);
-                }
-
-                @Override
-                public Void visitSpread(SpreadTree node, String matchExpression) {
-                    return fail(node, matchExpression);
-                }
-
-                @Override
-                public Void visitUnary(UnaryTree node, String matchExpression) {
-                    return fail(node, matchExpression);
-                }
-
-                @Override
-                public Void visitUnknown(Tree node, String matchExpression) {
-                    return fail(node, matchExpression);
-                }
-
-                @Override
-                public Void visitWhileLoop(WhileLoopTree node, String matchExpression) {
-                    return fail(node, matchExpression);
-                }
-
-                @Override
-                public Void visitWith(WithTree node, String matchExpression) {
-                    return fail(node, matchExpression);
-                }
-
-                @Override
-                public Void visitYield(YieldTree node, String matchExpression) {
-                    return fail(node, matchExpression);
-                }
-
-                @Override
-                public Void visitContinue(ContinueTree node, String matchExpression) {
-                    return fail(node, matchExpression);
-                }
-
-                @Override
-                public Void visitDebugger(DebuggerTree node, String matchExpression) {
-                    return fail(node, matchExpression);
-                }
-
-                @Override
-                public Void visitThrow(ThrowTree node, String matchExpression) {
-                    return fail(node, matchExpression);
-                }
-
-                @Override
-                public Void visitTry(TryTree node, String matchExpression) {
-                    return fail(node, matchExpression);
-                }
-            };
+    private final TreeVisitor<Void, String> treeVisitor = new MatchExpressionTreeVisitor();
 
     String validate(Rule rule) {
         CompilationUnitTree cut = parser.parse(rule.getName(), rule.getMatchExpression(), null);
